+++ src/ice/manage/dec/StreamingDecoder.cpp
@@ -15,13 +15,13 @@ StreamingDecoder::StreamingDecoder(std::string_view path,
 
 // 解码数据到缓冲区的接口
 double StreamingDecoder::decode(float** buffer, uint16_t num_channels,
-                                double start_frame, double frame_count) {
+                                size_t start_frame, size_t frame_count) {
     // TODO(xiang 2025-07-11): 实现流式解码
     return 0;
 }
 double StreamingDecoder::origin(
-    std::vector<std::span<const float>>& origin_data, double start_frame,
-    double frame_count) {
+    std::vector<std::span<const float>>& origin_data, size_t start_frame,
+    size_t frame_count) {
     return 0.;
 }
 }  // namespace ice
